[UPDATE] CAR GET POST API

+++ src/main/java/com/evdealer/ev_dealer_management/car/model/dto/dimension/DimensionPostDto.java
@@ -7,7 +7,7 @@ public record DimensionPostDto(
     Float widthFoldedIn,
     Float widthExtendedIn,
     Float heightIn,
-    Integer lengthIn,
+    Float lengthIn,
     Float wheelsSizeCm
 ) {
 }
